+++ tests/testthat/test-test_construction.R
@@ -1,6 +1,6 @@
 test_that("all learners can be constructed without custom args", {
   expect_silent(lrns(c(
-    "classif.glmnet",
+    "classif.cv_glmnet",
     "classif.kknn",
     "classif.lda",
     "classif.log_reg",
@@ -11,7 +11,7 @@ test_that("all learners can be constructed without custom args", {
     "classif.xgboost",
     "classif.multinom",
 
-    "regr.glmnet",
+    "regr.cv_glmnet",
     "regr.kknn",
     "regr.km",
     "regr.lm",